refactor(weather): Move view model layer to feature-specific directory
- Relocated all files from lib/view_model/ to lib/features/weather/view_model/
to better encapsulate state management within the weather feature.
- Updated all internal and external import paths accordingly.
