+++ lib/features/weather/view_model/city_search_state.freezed.dart
