+++ lib/features/weather/view_model/city_search_view_model.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../features/weather/repositories/firestore_weather_repository.dart';
-import '../features/weather/repositories/firestore_weather_repository_provider.dart';
+import '../repositories/firestore_weather_repository.dart';
+import '../repositories/firestore_weather_repository_provider.dart';
 import 'city_search_state.dart';
 import 'providers/app_router_provider.dart';
