+++ lib/features/weather/view_model/error_state.dart
@@ -1,6 +1,6 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-import '../core/network/api_error.dart';
+import '../../../core/network/api_error.dart';
 
 part 'error_state.freezed.dart';
