+++ lib/features/weather/view_model/error_state.freezed.dart
