+++ lib/features/weather/view_model/providers/city_name_validator_provider.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../features/weather/validator/city_name_validator.dart';
+import '../../validator/city_name_validator.dart';
 
 /// [CityNameValidator]を提供する[Provider]
 final cityNameValidatorProvider = Provider((ref) => CityNameValidator());
