+++ lib/features/weather/view_model/providers/city_weather_notifier_provider.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../core/network/response/result.dart';
-import '../../features/weather/models/forecast.dart';
+import '../../../../core/network/response/result.dart';
+import '../../models/forecast.dart';
 import '../city_weather_notifier.dart';
 
 /// [CityWeatherNotifier]を提供する[NotifierProvider]
