+++ lib/features/weather/view_model/weather_view_model_state.freezed.dart
